FEAN-30 셀러 권한 수정(등록 및 삭제) 기능 구현 - Sevice

+++ src/main/java/com/lotte/danuri/member/members/MemberService.java
@@ -9,7 +9,7 @@ public interface MemberService {
 
     long updateMemberInfo(MemberInfoReqDto dto);
 
-    int updateSellerAuth(SellerAuthReqDto dto);
+    long updateSellerAuth(SellerAuthReqDto dto);
 
     Optional<List<Member>> getSellers(SellerAuthReqDto dto);
